Move tb workflow out
