+++ src/nomad_simulations/schema_packages/workflow/__init__.py
@@ -1,4 +1,3 @@
-from .beyond_dft import BeyondDFTResults, DFTTBWorkflow
 from .equation_of_state import EquationOfState
 from .general import ParallelWorkflow, SerialWorkflow, SimulationWorkflow
 from .geometry_optimization import GeometryOptimization
@@ -7,4 +6,5 @@ from .molecular_dynamics import MolecularDynamics
 from .phonon import Phonon
 from .photon_polarization import PhotonPolarizationWorkflow
 from .single_point import SinglePoint
+from .tb import DFTTBWorkflow
 from .thermodynamics import Thermodynamics
